# Optimize `Resource::set_path()`