@Rindbee 2025-10-25
**Testing project:** [benchmark-resource-set-path.zip](https://github.com/user-attachments/files/23140543/benchmark-resource-set-path.zip)

## Benchmark

### Specifications

Linux Mint 22.2 (Zara) on X11 - X11 display driver, Multi-window, 1 monitor - Vulkan (Forward+) - integrated AMD Radeon Graphics (RADV PHOENIX) - AMD Ryzen 7 8745H w/ Radeon 780M Graphics (16 threads) - 25.22 GiB memory

Use a Resource array and set the resource paths of its elements to simulate the cached Resource.
Path format: `res://{idx}.res`.

Given that the number of cached items is limited in practice, the number of loop calls is increased to get a more accurate measurement.

### Case 0: Change the path to be empty

Sets the resource paths of all elements in the array to empty, and then resets the resource paths of all its elements again.

| times=5 | cache count | loop count | min cost | max cost | average cost |
| :----: | :-----: | :-----: | :----: | :----: | :----: |
| Before | 100 | 1000 | 47.9 | 49.7 | 48.8 (100%) |
| After | 100 | 1000 | 45.6 | 46.4 | 46.0 (94.3%) | 
| Before | 200 | 1000 | 89.7 | 94.3 | 92.9 (100%) |
| After | 200 | 1000 | 88.6 | 90.1 | 89.4 (96.2%) | 
| Before | 500 | 1000 | 223.7 | 234.1 | 230.0 (100%) |
| After | 500 | 1000 | 213.5 | 223.5 | 220.5 (95.9%) | 
| Before | 1000 | 1000 | 463.5 | 471.3 | 466.5 (100%) |
| After | 1000 | 1000 | 454.0 | 459.4 | 456.2 (97.8%) | 

### Case 1: Modify the path (not empty)

Change the resource paths of all elements.
Path format: `old_path` + `"test"`

| times=5 | cache count | loop count | min cost | max cost | average cost |
| :----: | :-----: | :-----: | :----: | :----: | :----: |
| Before | 100 | 1000 | 420.7 | 424.3 | 423.1 (100%) |
| After | 100 | 1000 | 420.8 | 423.8 | 422.0 (99.7%) | 
| Before | 200 | 1000 | 847.5 | 850.0 | 848.9 (100%) |
| After | 200 | 1000 | 846.5 | 850.2 | 848.1 (99.9%) | 
| Before | 500 | 1000 | 2149.6 | 2165.1 | 2155.2 (100%) |
| After | 500 | 1000 | 2148.9 | 2159.3 | 2155.6 (100%) | 
| Before | 1000 | 1000 | 4514.6 | 4542.0 | 4524.9 (100%) |
| After | 1000 | 1000 | 4502.0 | 4522.6 | 4514.7 (99.8%) |