@Rindbee 2025-11-10
**Testing project:** [benchmark-resource-set-path-03.zip](https://github.com/user-attachments/files/23443761/benchmark-resource-set-path-03.zip)

## Benchmark

### Specifications

Linux Mint 22.2 (Zara) on X11 - X11 display driver, Multi-window, 1 monitor - Vulkan (Forward+) - integrated AMD Radeon Graphics (RADV PHOENIX) - AMD Ryzen 7 8745H w/ Radeon 780M Graphics (16 threads) - 25.22 GiB memory

The following results were measured using an editor built with `scons`.

Use a Resource array and set the resource paths of its elements to simulate the cached Resource.
Path format: `res://{idx}.res`.

Given that the number of cached items is limited in practice, the number of loop calls is increased to get a more accurate measurement.

### Case 0: Change the path to be empty

Sets the resource paths of all elements in the array to empty, and then resets the resource paths of all its elements again.

| times=5 | cache count | loop count | min cost | max cost | average cost |
| :----: | :-----: | :-----: | :----: | :----: | :----: |
| Before | 100 | 10000 | 448ms | 459ms | 451ms (100%) |
| After | 100 | 10000 | 419ms | 422ms | 420ms (93.1%) | 
| Before | 200 | 10000 | 894ms | 934ms | 905ms (100%) |
| After | 200 | 10000 | 833ms | 847ms | 838ms (92.6%) | 
| Before | 500 | 10000 | 2275ms | 2292ms | 2285ms (100%) |
| After | 500 | 10000 | 2097ms | 2148ms | 2120ms (92.8%) | 
| Before | 1000 | 10000 | 4576ms | 4671ms | 4630ms (100%) |
| After | 1000 | 10000 | 4297ms | 4353ms | 4328ms (93.5%) | 


### Case 1-2: Modify the path (not empty)

Change the resource paths of all elements, and then resets the resource paths of all its elements again.
Path format: `old_path` + `"test"` 

| times=5 | cache count | loop count | min cost | max cost | average cost |
| :----: | :-----: | :-----: | :----: | :----: | :----: |
| Before | 100 | 10000 | 590ms | 599ms | 593ms (100%) |
| After | 100 | 10000 | 563ms | 573ms | 567ms (95.6%) | 
| Before | 200 | 10000 | 1205ms | 1211ms | 1207ms (100%) |
| After | 200 | 10000 | 1157ms | 1169ms | 1160ms (96.1%) | 
| Before | 500 | 10000 | 3233ms | 3340ms | 3265ms (100%) |
| After | 500 | 10000 | 3085ms | 3112ms | 3096ms (94.8%) | 
| Before | 1000 | 10000 | 6903ms | 6950ms | 6928ms (100%) |
| After | 1000 | 10000 | 6612ms | 6791ms | 6661ms (96.1%) |